test: add a new test module - `test_layout_element.py`
